+++ src/taggy-core/src/main/java/ch/fhnw/core/repository/TagsRepository.java
@@ -11,7 +11,7 @@ import ch.fhnw.core.domain.Tag;
 
 @Repository
 @Transactional
-public interface TagsRepository extends JpaRepository<Tag, Integer>{
+public interface TagsRepository extends JpaRepository<Tag, Long>{
 
 	Tag findByName(String tagName);
 	List<Tag> findByPictures(Picture picture);
