+++ src/taggy-core/src/main/java/ch/fhnw/core/services/PictureService.java
@@ -10,8 +10,8 @@ import ch.fhnw.core.domain.Tag;
 public interface PictureService {
 	
 	Stream<Picture> findByComment(String pictureComment);
-	Picture findById(Integer id);
-	Stream<Picture> findByTag_id(Integer id);
+	Picture findById(Long id);
+	Stream<Picture> findByTag_id(Long id);
 	Stream<Picture> findPictureByTag(Tag tag);
 	List<Picture> findPictureByTagsAnd(List<Tag> tags);
 	Stream<Picture> findPictureByTagsOr(List<Tag> tags);
